scenarios/openai_on_custom_dataset/ingest/requirements.txt
azure-core
openai
pandas
#azure-search-documents==11.4.0b9
azure-ai-formrecognizer
python-dotenv